+++ src/components/paginator/types.ts
@@ -1,4 +1,12 @@
 import { BaseElementProps, ElementPosition, ElementSize } from "@bbr/types";
+import { ButtonProps } from "@bbr/components/button";
+
+/** Configuration for the next buttons */
+export type NextButtonConfig = Omit<
+    ButtonProps,
+    | "onClick"
+    | "disabled"
+>;
 
 export type PaginatorProps = BaseElementProps & {
     /** Amount of pages */
@@ -19,7 +27,28 @@ export type PaginatorProps = BaseElementProps & {
     /** Size of paginator component elements */
     size?: ElementSize;
 
+    /** Configuration for next buttons */
+    nextButtonsConfig?: {
+        /** Configuration for the previous button */
+        previousButtonConfig: NextButtonConfig;
+
+        /** Configuration for the next button */
+        nextButtonConfig: NextButtonConfig;
+    };
+
     /** Display "Previous" \ "Next" buttons */
+    /**
+     * Display "Previous" \ "Next" buttons.
+     * @deprecated Use `nextButtonsConfig` prop instead
+     * @example
+     * <Paginator
+     *   ...
+     *   nextButtonsConfig={{
+     *     previousButtonConfig: { ... },
+     *     nextButtonConfig: { ... }
+     *   }}
+     * />
+     */
     showNextButtons?: boolean;
 
     /**
@@ -28,29 +57,36 @@ export type PaginatorProps = BaseElementProps & {
     */
     nearPagesCount?: number;
 
-    /** String resources for component */
+    /**
+     * Resources for paginator component. Used for buttons captions and titles.
+     * If not set, default values will be used.
+     */
     resources?: {
         /**
          * Previous page button caption
          * @default "Previous"
+         * @deprecated Use `nextButtonsConfig.previousButtonConfig.caption` prop instead
          */
         previousPageCaption?: string;
 
         /**
          * Previous page button title.
          * Visible only if set
+         * @deprecated Use `nextButtonsConfig.previousButtonConfig.title` prop instead
          */
         previousPageTitle?: string;
 
         /**
          * Next page button caption
          * @default "Next page"
+         * @deprecated Use `nextButtonsConfig.nextButtonConfig.caption` prop instead
          */
         nextPageCaption?: string;
 
         /**
          * Next page button title.
          * Visible only if set
+         * @deprecated Use `nextButtonsConfig.nextButtonConfig.title` prop instead
          */
         nextPageTitle?: string;